Part 2

+++ Day_6/Race.cs
@@ -2,6 +2,6 @@ namespace Day_6;
 
 public struct Race
 {
-    public int Time;
-    public int Distance;
+    public long Time;
+    public long Distance;
 }
